feat: Ghetto-pytest refactor (aka compilation-station)
A bit liberal in that commit summary, but it's not entirely wrong.

We don't yet have support for pytest, which was the original plan. So
instead I've re-engineered Lennart's solution, but in python (since I
couldn't get his solution to work).

What we get is a python script which runs all the tests found in ./tests
(or as specified as TEST_DIR). It also tracks success/failures, so it
does the same things as pytest, but bare minimum for our usecase.