feat: Added <api>/check/<test> and added run-all-tests-via-api.py
Although the <api>/check endpoint worked just fine locally, it times out
the workload on edge. So we need to be able to call each test one by
one.

So in addition to this, I added a <api>/list endpoint which lists all
tests, then a /check/<test> endpoint which calls a singular test. This
allows us to run just a specific test, if need be.

Also added instructions on how this works to README.md